GW-71: Refactored folder structure
