+++ client/src/pages/about/components/OurTeam.jsx
@@ -1,0 +1,31 @@
+import React from 'react';
+import { Box } from '@mui/material';
+import GridContainer from 'shared/components/GridContainer';
+import { Typography, DropdownMenu } from 'shared/components/index';
+import TeamButtons from './TeamButtons';
+import TeamGridContent from './TeamGridContent';
+
+export default function OurTeam({ data, states }) {
+	// eslint-disable-next-line prettier/prettier
+	const {
+		selectedYear, selectedButton, selectedCard,
+		setSelectedYear, setSelectedButton, setSelectedCard,
+	} = states;
+
+	return (
+		<>
+			<Box className='w-full py-4 sm:py-6 3xl:py-10'>
+				<Typography variant='h1' text='Meet our team.' />
+			</Box>
+			<Box>
+				<Box className='flex flex-col'>
+					<TeamButtons selectedButton={selectedButton} setSelectedButton={setSelectedButton} />
+					<DropdownMenu selectedYear={selectedYear} setSelectedYear={setSelectedYear} source='About' />
+				</Box>
+				<GridContainer className='sm:mt-14'>
+					<TeamGridContent data={data} selectedCard={selectedCard} setSelectedCard={setSelectedCard} />
+				</GridContainer>
+			</Box>
+		</>
+	);
+}
