+++ client/src/pages/about/components/TeamGridContent.jsx
