+++ client/src/pages/about/components/TeamGridItem.css
