+++ client/src/pages/about/components/TeamGridItem.jsx
